Componente checkout finalizado

+++ src/app/api/admin/delete-product/[id]/route.js
@@ -1,4 +1,5 @@
 import conectToDB from "@/database";
+import AuthUser from "@/middleware/AuthUser";
 import Product from "@/models/product";
 import { useRouter } from "next/navigation";
 import { NextResponse } from "next/server";
@@ -8,25 +9,30 @@ import { NextResponse } from "next/server";
 export const dynamic = 'force-dynamic';
 
 
-export async function DELETE(req){ 
-    //console.log(req)
-    //const router = useRouter();
-    //console.log(router.query);
+export async function DELETE(req, {params}){ 
+    
     try {       
         await conectToDB();
-        const { searchParams, pathname } = new URL(req.url);
-        //const id = searchParams.get('id');
-        const id2 = pathname.slice(26);
-        
+        const isAuthUser = AuthUser(req);
 
-        if(!id2) return NextResponse.json({ success : false, message : 'ID del producto no encontrado' });
-        
-        const deletedProduct = await Product.findByIdAndDelete(id2);
-        if(deletedProduct){
-            return NextResponse.json({ success : true, message : 'Producto eliminado correctamente' });
-        }else{
-            return NextResponse.json({ success : false, message : 'Error al borrar el producto' });
+        if(isAuthUser?.role === 'admin'){
+            const id2 = params.id;
+            
+
+            if(!id2) return NextResponse.json({ success : false, message : 'ID del producto no encontrado' });
+            
+            const deletedProduct = await Product.findByIdAndDelete(id2);
+            if(deletedProduct){
+                return NextResponse.json({ success : true, message : 'Producto eliminado correctamente' });
+            }else{
+                return NextResponse.json({ success : false, message : 'Error al borrar el producto' });
+            }
+
+        } else{
+            return NextResponse.json({ success : false, message : 'No estas autenticado' });
+
         }
+        
     } catch (error) {
         console.log(error);
         return NextResponse.json({
